UNIC: Debugfs supports query of ip and guid table's list and specification
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I850RQ
CVE: NA

-----------------------------------------------------------------

IP table and GUID table specification including function number,
address and state of the table which have been configured to
hardware is significant for user to choose the scheme of using
the table.

This patch is to use debugfs to query the specification and list
of the IP and GUID table.

Signed-off-by: Haibin Lu <luhaibin10@hisilicon.com>
Signed-off-by: Junxin Chen <chenjunxin1@huawei.com>