net: hns3: fix wrong page_region param for 8521 phy custom interface
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I7YRUW
CVE: NA

----------------------------------------------------------------------

For the 8521 phy, the page_region parameter is not required. Therefore,
the value of the page_region parameter is fixed to 0
when nic_get_phy_reg is called.
However, the page_region parameter in the macro function is not deleted.

This patch fixes this bug by removing the page_rea parameter
from the macro function.

Signed-off-by: Jijie Shao <shaojijie@huawei.com>