Support configuration of fastpath feature
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/I99GLS
CVE: NA

--------------------------------------------------------

Currently, there is a limitation that fastpath can only
be used in single function on the chip. In order to obtain
the maximum delay benefit under the chip limitation, this
patch adds fastpath configuration in different scenarios,
mainly including the following:
1. Fastpath is turned on when the driver is loaded
and turned off when the driver is unloaded.
2. Fastpath is turned off when SRIOV is enabled and
turned on when SRIOV is disabled. At the same time,
before configuring fastpath in this scenario, it is
necessary to stop the flow to avoid abnormal
problems in the chip.

Signed-off-by: Zihao Sheng <shengzihao1@huawei.com>